pyproject.toml
[project]
name = "adl1t"
version = "1.0"
description = "Anomaly detection at the level 1 CMS trigger."
authors = [
    {name = "Patrick Odagiu",email = "podagiu@student.ethz.ch"},
    {name = "Victor Jimenez",email = "vjimenez@ethz.ch"}
]
license = {text = "MIT"}
readme = "README.md"

requires-python = ">=3.10"
dependencies = [
    "adl1t-datamaker @ git+https://github.com/bb511/adl1t_datamaker.git@master",
    "torch (==2.*)",
    "torchvision (==0.*)",
    "pytorch-lightning (==2.*)",
    "torchmetrics (==1.*)",
    "hydra-core (==1.*)",
    "rich (==13.*)",
    "pre-commit (==3.*)",
    "pytest (==7.*)",
    "ipdb (>=0.13.13,<0.14.0)",
    "mlflow (==3.1.0)",
    "hydra-optuna-sweeper (>=1.2.0,<2.0.0)",
    "hydra-colorlog>=1.2.0,<2.0.0",
    "rootutils (>=1.0.7,<2.0.0)",
    "numexpr (>=2.14.1,<3.0.0)",
    "pathvalidate (>=3.3.1,<4.0.0)",
    "hydra-submitit-launcher (>=1.2.0,<2.0.0)",
    "colorama (>=0.4.6,<0.5.0)",
    "mplhep>=0.4.1",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[project.optional-dependencies]
quant = [
    "keras (<3.13)",
    "hgq2 (>=0.1.5,<0.2.0)",
    "hls4ml @ git+https://github.com/fastmachinelearning/hls4ml.git",
]
cap = [
    "capmetric @ git+ssh://git@github.com/viictorjimenezzz/cap-metric.git@main",
]

[tool.pytest.ini_options]
addopts = [
  "--color=yes",
  "--durations=0",
  "--strict-markers",
  "--doctest-modules",
]
filterwarnings = [
  "ignore::DeprecationWarning",
  "ignore::UserWarning",
]
log_cli = "True"
markers = [
  "slow: slow tests",
]
minversion = "6.0"
testpaths = "tests/"

[tool.coverage.report]
exclude_lines = [
    "pragma: nocover",
    "raise NotImplementedError",
    "raise NotImplementedError()",
    "if __name__ == .__main__.:",
]
